requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
sgp4>=2.22
apscheduler>=3.10
python-dotenv>=1.0
